backend/src/test/java/com/asdc/dalexchange/service/impl/SoldItemServiceImplTest.java
package com.asdc.dalexchange.service.impl;
import com.asdc.dalexchange.dto.SoldItemDTO;
import com.asdc.dalexchange.mappers.impl.SoldItemMapperImpl;
import com.asdc.dalexchange.model.SoldItem;
import com.asdc.dalexchange.repository.SoldItemRepository;
import com.asdc.dalexchange.specifications.SoldItemSpecification;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;

import java.util.ArrayList;
import java.util.List;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.mockito.Mockito.*;

public class SoldItemServiceImplTest {

    @Mock
    private SoldItemRepository soldItemRepository;

    @Mock
    private SoldItemMapperImpl soldItemMapper;

    @InjectMocks
    private SoldItemServiceImpl soldItemService;

    @BeforeEach
    void setUp() {
        MockitoAnnotations.initMocks(this);
    }

    @Test
    void testGetallSoldProduct_Success() {
        // Mock data
        Long userId = 1L;
        List<SoldItem> mockSoldItems = createMockSoldItems(userId);
        List<SoldItemDTO> mockSoldItemDTOs = createMockSoldItemDTOs(userId);

        // Mock repository behavior
        when(soldItemRepository.findAll(SoldItemSpecification.bySellerUserId(userId))).thenReturn(mockSoldItems);

        // Mock mapper behavior
        when(soldItemMapper.mapTo(any(SoldItem.class))).thenAnswer(invocation -> {
            SoldItem soldItem = invocation.getArgument(0);
            return createMockSoldItemDTO(soldItem);
        });

        // Call the service method
        List<SoldItemDTO> soldItemDTOs = soldItemService.GetallSoldProduct(userId);

        // Assertions
        assertEquals(mockSoldItemDTOs.size(), soldItemDTOs.size());
        // Add more assertions based on your DTO and expected behavior
    }

    // Helper method to create mock sold items for testing
    private List<SoldItem> createMockSoldItems(Long userId) {
        List<SoldItem> soldItems = new ArrayList<>();
        // Create mock SoldItem objects
        return soldItems;
    }

    // Helper method to create mock SoldItemDTO objects for testing
    private List<SoldItemDTO> createMockSoldItemDTOs(Long userId) {
        List<SoldItemDTO> soldItemDTOs = new ArrayList<>();
        // Create mock SoldItemDTO objects
        return soldItemDTOs;
    }

    // Helper method to create a mock SoldItemDTO based on SoldItem for testing
    private SoldItemDTO createMockSoldItemDTO(SoldItem soldItem) {
        SoldItemDTO soldItemDTO = new SoldItemDTO();
        // Map attributes from SoldItem to SoldItemDTO
        return soldItemDTO;
    }

    // Add more test cases to cover edge cases and other methods as necessary

}
